2d Minkowski mostly working
When the 2d minkowski was enabled the nef2p2 was bailing out too early.

I added a simple p2nef2 based on what projection.cc was doing, and
applied it to the outer_boundary of the polygon_with_holes_2. It now
seems to work for simple scad code such as:

minkowski() {
        translate([25,0,0])  square(20);
    circle(r=3);
}