Revert "Output LWPOLYLINE objects instead of LINE for DXF export. Patch submitted by Lars Kruse"
This reverts commit 060a6c654de5ebd14ccb889881fa6c4d3fb9d763.
The reason is that the output DXF won't import in a number of applications due to a malformed DXF (AutoCAD, Solidworks). The correct fix is probably to use a DXF library to generate a correct DXF. This is only a case for more recent (R14+) versions of DXF, e.g. to be able to use the LWPOLYLINE entity.