Make OpenSCAD look better on retina displays.
Apps need an NSPrincipalClass entry in Info.plist to look good on retina
displays, see "Framework-Scaled Mode Provides Automatic Scaling" in
http://developer.apple.com/library/mac/#documentation/GraphicsAnimation/Conceptual/HighResolutionOSX/Explained/Explained.html

It's also what we did in Chromium ( https://codereview.chromium.org/10069029 )
and what was suggested in openscad issue #279.

This only makes the window decoration and all text pretty. The 3d view
is still pixel-doubled: OpenGL is a pixel-based API and needs an
explicit opt-in. This can be addressed in a future patch.

While here, also consistently use tabs instead of spaces.