In this commit i'll change just the Qt project, c and header files to get OpenSCAD building and working on Windows 7 compiled with Visual Studio Express 2012.
A couple of things needed changing :
* MSVC 2012 has no trunc or round, i have added those to mathc99 and included mathc99 to the files needed.
The implementations were taken from boost.
* MSVC 2012 stl does now allow accessing out of bounds elements in std:vector as an array, so "chain->objects[i]" will
fail in "OpenCSGRenderer::renderCSGChain". Whatever the platform, it will me a good idea to change this
code to something that does not trigger undefined behavior.
* MSVC 2012 cannot figure out the types involved in "if (file_size ==  80 + 4 + 50*facenum) {", i have added
a cast to force it to "if (file_size ==  static_cast<std::streamoff>(80 + 4 + 50*facenum)) {".

The next round would be regression testing and debugging and possibly more changes to the OpenSCAD main code.

The round after that would be changing the .pro and .pri files to create correct visual studio solutions.