Fix warning - Failed to open file OPENSCAD_DIRECTORY: file to open is a directory
The issue appears to be that an empty entry is put into inputFiles when
no others are provided.

    if (!inputFiles.size()) inputFiles.push_back("");

and then that empty entry is used in an assembled path with a path

    MainWindow *m = new MainWindow(assemblePath(original_path,
inputFiles[0]));

which makes it appears like a real file, which it is not.

A fix is:
Testing if the input file 0 is empty and if empty using the empty string
-- else if not empty assembling as before.

Also removed a dead variable called filename.