Try to keep cursor and scroll in place on refresh
When the file is saving, the cursor position was resetting to the top and you should keep
scrolling by hand every time you save the file.

This patch saves the scroll and cursor position and set it again after the file is refreshed so
we can continue edition from where we were.