write output to temporary file, commit only after completion
cmdline() is modified to write into <output_file>~ and only after
succesful completion of the given command rename this file into
<output_file>. In case of failure the file is deleted. This prevents
dangling errorneous output files and overwriting valid (but old)
files with new trash.

This patch is required for robust use of OpenSCAD in Makefiles. In
case of a crash the output file must not be touched. Otherwise the
update file date will trick Makefile into assuming a succesful build.