Move translation files to locale folder instead of po.
While this is not the standard name for the files, this makes the
resource handling easier as the PlatformUtils code already handles
lookup for both installed and dev layouts. At least for Linux the
installed name would be /usr/share/locale, so we use that in source
too for now.