Add optional menu driven scaling tics to the large axes
Tim Deagan 12/26/2014

While add-on rulers can be extremely useful, there are many times that a quick scaling marker would be handy. This feature generates tics every 10 units, driven off the log10 of the l variable defining axes length.  As you zoom in or out, the tics automatically re-scale themselves.  Every tenth tic is slightly larger.  A menu item is added to the view menu to enable or disable the feature.

I experimented with menu driven scaling (.01, .1, 1, 10, etc.), but using small increments when zoomed out brought the app to it's knees.  Adding a visual indicator of the current scaling might be nice, but I'm nto sure where to put it (possibly below the small axes,) since it changes as zooming and would clutter the console.

 Changes to be committed:
	new file:   images/scalemarkers.png
	modified:   openscad.qrc
	modified:   src/GLView.cc
	modified:   src/GLView.h
	modified:   src/MainWindow.h
	modified:   src/MainWindow.ui
	modified:   src/QGLView.h
	modified:   src/mainwin.cc