fix building of remaining binaries
the previous patch has removed the -DENABLE_EXPERIMENTAL flag, assuming
that it is not required any more without openscad_nogui being built.
while that is sufficient for the test suite, other binaries built in the
tests/ directory would still use those flags.

in order to make them buildable again, the -DOPENSCAD_NOGUI compile time
option is added (as the resulting binaries are command-line-only tools,
which was previously implicit by setting -DOPENSCAD_TESTING), and
-DENABLE_EXPERIMENTAL is restored.