interactive number adjust
When the cursor is on number, you can increment or decrement it with Alt-Up or Alt-Down keys
and see the change of model in real time.
The Alt-Left and Alt-Right keys navigates to left or right within the number and allow you to change
the increment value. Navigating to right over the number, new decimal digits are added automatically.
When the update of preview is slow and you change the number quickly, the frame rate is limited, however
when you stop to increment, the actual model is shown as soon as possible.