Add echo() expression.
This is mainly for debugging expressions or functions as the existing
echo() is only usable on statement level.
The echo() expression works like let(), including the sequential assignment
with the additional side effect of printing all the arguments and if there
is an expression following the echo(), this expression is evaluated and
printed last.