Fix issue 214. When a syntax error is reported, we report the source file name and the line number within the source file where the error occurred. This works even if the syntax error is within an included file.
The parse() function, and the places where it is called, have been changed
as necessary to enforce the convention that the second argument is the source
filename, and not the name of the parent directory containing the source file.
Previously, different calls to parse() made different assumptions about
the meaning of the second argument, which probably means include and import
were broken in some cases.