Fix window visibility handling.
The issue is caused by restoreState() calling setVisible() on the Dock
with the old serialized values which triggers the automatic settings
handling to write that state back into the Qt settings. Solution for now
is to fetch the current state from Qt settings and restore those after
the restoreState() call.