add 'circulate' to linear_extrude
This integral value may be specified in the case where the extrusion is
of an "essentially circular", equal-angle-sampled and origin-centered
shape, such a thread profile.  It allows control over which points in
the previous slice are matched up with points in this slice, and must
be closely coupled with the profile's angular sampling, the twist count
and number of slices to give an aesthetic result.

When circulate= is not specified, no behavioral change is intended.