Fixed bad axis labeling at small zoom levels. #2020
There was an issue that caused the labels to be closely bunched up at
small zoom levels. This was because the axis scale calculation was
incorrect. We adjust the scale by a whenever the viewport changes in
size by a factor of ten by doing:

scale_length = 10**truncate(log10(viewport_size))

truncate(0.99) and truncate(-0.99) and everything in-between map to
zero, giving us a factor of 100 before the scale changes, when it
should instead be a factor of 10. Using floor() instead of truncate()
fixes the problem.