Feature addition: When auto reload happens, the main window is raise()d to the front. This is very useful with generating the .scad file from another application that would cover most of the screen space. I use openscad together with inkscape and inkscape likes to be in front most of the time...
Added scripts/docker-build.sh -- a quick wrapper to build on linux
bringing all prerequisites into a docker container to not pollute the build host.
Read and adjust the script before use, some things are hardcoded there.
It works with a git checkout outside the container. So that all object files and
the binary are easily accessible.
It shortcuts to the next needed step but is quite slow doing so.