Coverity: CID 25105: Uninitialized scalar field (UNINIT_CTOR).
Non-static class member width is not initialized in this constructor
nor in any functions that it calls.
Non-static class member height is not initialized in this constructor
nor in any functions that it calls.
Non-static class member aspectratio is not initialized in this constructor
nor in any functions that it calls.
Non-static class member orthomode is not initialized in this constructor
nor in any functions that it calls.