Coverity: CID 25062: Uninitialized pointer field (UNINIT_CTOR).
Non-static class member t is not initialized in this constructor nor in
any functions that it calls.
Non-static class member c is not initialized in this constructor nor in
any functions that it calls.
Non-static class member p is not initialized in this constructor nor in
any functions that it calls.
Non-static class member n is not initialized in this constructor nor in
any functions that it calls.