Coverity: CID 25112: Uninitialized scalar field (UNINIT_CTOR).
Non-static class member center is not initialized in this constructor nor
in any functions that it calls.
Non-static class member invert is not initialized in this constructor nor
in any functions that it calls.
Non-static class member convexity is not initialized in this constructor
nor in any functions that it calls.