Coverity: CID 188637: Uninitialized scalar field (UNINIT_CTOR).
Non-static class member width is not initialized in this constructor nor
in any functions that it calls.
Non-static class member height is not initialized in this constructor nor
in any functions that it calls.
Non-static class member rx is not initialized in this constructor nor in
any functions that it calls.
Non-static class member ry is not initialized in this constructor nor in
any functions that it calls.