Coverity: CID 188642: Uninitialized pointer field (UNINIT_CTOR).
Non-static class member polySet is not initialized in this constructor
nor in any functions that it calls.
Non-static class member x is not initialized in this constructor nor in
any functions that it calls.
Non-static class member y is not initialized in this constructor nor in
any functions that it calls.
Non-static class member z is not initialized in this constructor nor in
any functions that it calls.
Non-static class member idx_v1 is not initialized in this constructor nor
in any functions that it calls.
Non-static class member idx_v2 is not initialized in this constructor nor
in any functions that it calls.
Non-static class member idx_v3 is not initialized in this constructor nor
in any functions that it calls.