Make for() UNION optional
This adds a special case in for-loop arguments that prevents the
implied union of its children.

   for(..., union=false)

The children are then passed to the parent node as-is during the
parse phase, rather than trying to juggle things in GeometryBuilder.

This is a small step toward the more expansive feature that removes
the default implied unions, as proposed in #350.  It could also be
followed up with a minimally-invasive refactor of intersection_for
to simplify the code.