Fix missing warning about double assignment
Ensure that mainFilePath is an absolute path during parsing.

The logic for generating warnings about double assignment generally
seems to assume relative paths, and in fact mainFilePath is made
absolute in the recursive parse() call for a use <...> statement. So
if mainFilePath is not absolute from the start the code gets confused.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>