Restore Option+Arrow behavior on Mac.
As noted in [#2885](https://github.com/openscad/openscad/issues/2885),
Option+Arrow keys don't work in OpenSCAD on OS X anymore.  This updates
OpenSCAD so that Alt+Shift (AKA Option+Shift) is the modifier to use
the number precision extension that is bound to Alt+Left and Alt+Right
on other OSes.