Work-around for test failure at high -jN
When running the testsuite against an Xvfb $DISPLAY with -jN, there
are occasional failures with "Unable to open a connection to the X
server". This appears to be a bug in Xvfb with rapid parallel connects.

This patch implements a work-around: The testsuite sets the
environment variable OPENSCAD_X_CONN_RETRY; this causes openscad to
retry the X connection in case of failure, which avoids the test
failure.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>