Better fix for test failures at high -jN
Tests fail occasionally at high -jN with a failure to connect to the X
server on Xvfb. By default, the X server resets itself when all
clients disconnect, and this leaves a window where new connects can
fail during the reset. This can be avoided by passing the -noreset
flag to the X server.

This replaces the previous https://github.com/openscad/openscad/pull/3024
patch to solve the same problem.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>