Work-around for test failures on mips
The llvmpipe software-renderer of Mesa seems not stable on the MIPS
architecture. There are numerous crashes in the test suite when using
llvmpipe, as seen in the Debian build farm.

This patch implements the similar work-around that already exists for
PowerPC, to use the softpipe renderer (which was significantly faster
than llvmpipe anyway on MIPS in my tests).

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>