Implement lexical scoping for function literals.
The central change is to move the dynamic scoping handling out of the
lifetime of the Context class. Context and all derived classes are now
handled via shared pointer so they can hang around inside a function
definition as closure representing the static scoping stack.
The dynamic part is moved into ContextHandle which is supposed to be
created and destroyed as the evaluation traverses the expression tree.
The ContextHandle object calls push()/pop() on it's associated Context
on construction/destruction.