Fix winding order broken by 2D-in-3D optimization
Previous commit attempted to restore counter-clockwise winding order
on 2D-in-3D Polygons, but the real problem was that the 2D shapes
should never have been inverted.  I had optimized the transform3d
method to perform a 2d transform if the supplied matrix was purely
2d, but that skipped over the check for whether the winding order
needed swapped.  Once I re-added that check, the previous commit
was no longer necessary.  (though it cleans up the code a bit so
I'm leaving it there)