InputEventMapper: Break out generateDeferredEvents method
We are going to want to do something more sophisticated here.  In
particular, we are going to want to know, after we have looked at all
the axis values, whether these axis values warranted generating any
events - since, if they didn't, we can stop calling this code until
one of the axis values changes.

In this commit, therefore, put the event generation into its own
function.  This will make things a bit more convenient in the future
and also a bit clearer.

The return value is not used yet.

No functional change.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>