InputEventMapper: Disable timer when not needed
This 30ms timer is only needed when the joystick axis position implies
continuous periodic event generation.  At other times it can be
disabled.

This is important (particularly for users without joysticks) because
otherwise this 30ms timer causes an idle openscad to waste a
considerable amount of cpu, draining laptop batteries etc.

So arrange to have the timer running only when needed.

We enable the timer any time we change any of the things that might
affect event generation by generateDeferredEvents.  If that doesn't
result in the generation of any actual events, we disable it again.

We do the joystick position "scrollbar" update as before, but again,
once we have done it, there is no need to do it again.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>