Fix uncaught exception from boost::fs
The do_export() function wants to switch to the parent directory of
the .scad source file for .ast and .csg output. But the path to that
directory is computed at the wrong place, when a different directory
is the current one. This causes failure when the source file is
specified by relative path.

Fixes #3579

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>