Fix export_import_pngtest.py for relative source path
Use an absolute destination path for .csg output. This is necessary to
put the destination file in the correct place when the source path is
a relative path.

When outputting .csg, openscad interprets the destination path
relative to the parent directory of the source file, not relative to
the current directory when the program was started. This means that
something like `openscad a/b/c.scad -o a/b/c.csg` will try to write
a/b/a/b/c.csg, not a/b/c.csg as was expected by
export_import_pngtest.py.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>