Lookup tail recursive function calls independently for each call.
The old tail recursion logic would assume that a call to a function with the
same name is therefore a tail call. This breaks on examples like:
  function f() = let(f = function() 5) f();
The fix is to perform a context-based function lookup for tail call candidates,
rather than using a name comparison.

As a fortunate side effect, this also enables tail recursion optimizations
between calls to unrelated functions.

Fixes issue #3730.