support --hardwarnings for PNG and ECHO output (#3616)
As reported in #3616, --hardwarnings didn't always cause openscad to
exit with a non-zero error code when a warning was encountered. It did
for STL (and likely other geometry outputs) because those exports exited
with an error code if the root node was empty, but PNG and ECHO don't
check that.

I thought that relying on the geometry check seemed a little accidental,
so the fix I've made is to prevent the file context initialization from
swallowing the HardWarningsException.

I also fixed the hardwarnings test. The test passed only because it was
running the openscad executable with invalid commandline parameters.
OpenSCAD printed usage and exited with an error. The SUFFIX parameter is
not supported by the Cmake add_failing_test function, and shouldfail.py
does not automatically add a - for stdout output.