Remove customizer support for unconstrained vector variables.
Existing customizer support for vector variables is good for 1-4 element vectors of numbers, and bad for all other vectors.
Other vectors could be edited with a text editor only, which would then be parsed in an unreliable and unpredictable way, causing plenty of undesirable behavior.
Even if this would be replaced with a better parser, the limited compatibility between 1-4-number vectors and unconstrained vectors would be a problem.
Given that the makerbot customizer doesn't support vectors anyway, this patch removes support for unconstrained vectors entirely, and leaves only 1-4-number vectors.