Build most recently modified files first
This helps fail builds fast (and can pick which files to build first by just touching them), esp. when many files need to be rebuilt (e.g. when modifying a core header often but mostly caring about the compilation of a couple of C++ files).

Disabled on Windows (forks out to bash script tested on Mac and Linux), or if -DSORT_BUILD=0 passed to cmake