Fix error handling when creating an offscreen OpenGL context on Windows.
Fixes #4180

When animating from the command line interface, the code registers the
window class on each frame. This worked for the first frame, but the
registration for the second frame seemed to fail. Actually,
RegisterClassW was merely indicating that the second registration was
redundant.

This change ignores ERROR_CLASS_ALREADY_EXISTS, which seems to be the
least intrusive way to solve the problem.

I tested the solution in a proof-of-concept program, but I haven't tested
this actual patch because building OpenSCAD for Windows is onerous.

(There will probably still be bugs here.  It looks like the code never
destroys the windows it creates, and it creates a new one for each frame.
And each of those windows has its own DC because of the CS_OWNDC in the
window class.  So a many-frame animation is going to burn up GDI
resources.  Those lingering issues will be easier to detect and debug
once we can get past the initialization issues.)