Drafted implementation of a two pass rendering in OpenCSGRenderer
When using VBO and shader (eg: when using edge visualization or doing a
mouse selection) there is depth tests failures because of small calculation difference between
difference between the fixed pipeline vertex processing used in OpenCSG and
the shader based pipline.

To fix it, this PR implement a two pass algorithm.