In implementation compilation unit: move corresponding header to top.
Move header foo.h within implementation foo.cc to the top. That ensures
that the header at least is self-contained.

This revealed, that the core/SurfaceNode.h header was missing an
include of core/Value.h.

This PR leaves out changing src/glview, as there are
too many header dependencies that break, and should be addressed
separately.