Don't #define _USE_MATH_DEFINES in *.cc files but as compile define.
Setting the #define in the compilation unit makes it dependent if
the same header had been included before; if <cmath> has been included
before (and we don't know that, as we don't know what system headers
do between each other), the include-guard will prevent our definition
to ever been seen.

So the only way to make sure <cmath> sees the _USE_MATH_DEFINES
define is to set it on the compile command line, i.e. in the
CMakeLists.txt