FIX issue #5318
In issue #5318 was reparted crash when user was using the right click/jump-to definition feature recently implemented in openSCAD.
The crash only happens when jumping-to definition that is located in a not yet loaded file.

The crash is caused by the fact that when opening a new file, the existing node tree is detroyed
and replaced with a new one from the opened file.

In this PR, we fixed the issue by separating the opening of a new file in the editors panel
and loading the tree that contains it. To make it non breaking API, this was one by adding an extra parameter
with default value to the loading functions

This fix the issue.

Signed-off-by: Damien Marchal <damien.marchal@univ-lille1.fr>