Allow to do --animate work in shards.
Introduce new `--animate_sharding=<shard>/<num_shards>` flag that
allows to output a subset of animation frames.
e.g. `--animate_sharding=2/5` only outputs the second 1/5 of all shards.

This allows an easy parallelization on a machine with multiple cores,
or even distribute the work to multiple machines.

A typical use could be like this: run two instances and divide
the work into two shards that output the first and second half
of the frames.

```bash
   openscad -q --animate=42 --animate_sharding=1/2 foo.scad &
   openscad -q --animate=42 --animate_sharding=2/2 foo.scad &
   wait
```

(of course, one would use probably more shards closer to the
number of cores on the machine).

Issues: #5467