[UX] Add of the "Jump to..." navigation method and clean the "Window" menu of its duplicated elements (#5630)
* Add of the "Jump to..." navigation method and clean the "Window" menu of its duplicated elements
* Add overlay hint to indicate to users the selected dock.
* Move the title management on topLevelChanged from MainWindow into Dock.
* Replace the dock title management in MainWindow & TabManager by the one in Dock.

The MainWindow class was having a lot of slots and method to handle topLevel changes.
eg:
   void consoleTopLevelChanged(bool);

There was several duplication in the code, with one slot per dock.
Most of the code was there to handle the two version of the dock's title.

Now that the title is implemented in Dock  class, all that code is not needed anymore
So the PR remove it.

The PR also move into dedicated method some code fragment used in several places.
This is why the following method appears:
- MainWindow::getCurrentFileName();

In TabManager there was also direct call to the MainWindow method, including
slot regarding topLevel changes. see TabManager::getTabContent() for an example.
I see no good reason justifying the cross dependency between TabManager and MainWindow, so
all the direct calling to MainWindows methods have been moved into the MainWindow::onTabManagerEditorChanged slot
The slot being connected to the new signal TabManager::currentEditorChanged

Finally all connection between tabmanager and maindows have been moved from TabManager to MainWindow.
Ideally the TabManager class should't include "MainWindow.h".

* Fix keyboard navigation with CTRL+K and CTRL+L problem.

The problem exists in master branch of openscad and is related to how we select the docks by using only
if the dock has focus or not. This seems to be not enough, and we also need to take the window
activation state.

---------

Signed-off-by: Damien Marchal <damien.marchal@univ-lille1.fr>
Co-authored-by: Chris Mayo <aklhfex@gmail.com>