Fix issue #5695
The implemention is now more unified with an effort in factoring code.

The rubberband states are the following:
- when a window menu entry (next/prev or jump to) is hovered, the rubberband indicates which dock will be the one activated
- when a window menu entry (next/prev or jump to) is triggered, the dock is ativated and the rubbedband hides.

The shortcut Ctrl+K and Ctrl+H have been implemented differently so we differentiate
when a next/prev dock is activated/hovered because of a keyboard shortcut or because of hovering
the over a menu entry.