Clean MainWindow show/hide action (#5706)
* Fix minor typo in MainWindow::onWindowActionSelectEditor

* Adds comments in MainWindow.h

* Remove the windowAction in mainwindow.ui because they are not needed anymore

In PR #5630 the menuWindow behavior was changed to remove the "quasi duplicated" menu entries.
- the windowActionHideXXX actions are now using the qt default dock action
- the windowActionSelectXXX actions are now using the Jump-To menu.

So both windowActionHideXXX and windowActionSelectXXX can be removed.
Specific behavior when showing/hiding will be implemented using dedicated slot connected to visibilityChanged of the doc.

* Remove action in Dock.

The setAction is just there to allow the synchronization of the
action->setChecked() status to reflect the visibility one of the Dock.

This is not needed anymore because of the use of the Qt default dock action.

* Insure that saving the settings stills work and apply beautify.sh