Fixup VENV created from inside an AppImage.
When created from the AppImage, the VENV points to the temporary mount
point of the AppImage, e.g. /tmp/.mount_OpenSCCpPaio - that is obviously
no good for any later runs.

To fix that, we point the link to the magic /proc/self/exe so it can
always just call itself as the python interpreter.