Add FLAG_HIGHGLITH_SELECTED, FLAG_HIGHLIGHT_IMPACTE and BACKGROUND in CSGNode
To increase the amont of visual feedback when an object is selected we add more highlighting states.
HIGHLIGHT represents the object selected in source code with the # (dash) operator
HIGHLIGHT_SELECTED represents the object selected by mouse selection in the 3D view.
HIGHTLIGHT_IMPACTED represents the objects that will be impacted if the one selected is changed.
HIGHTLIGHT_BACKGROUND is there for moving the background management as a subcase of HIGHLIGHT.