Make the highlighted product an independant parameter of the OpenCSGRenderer
To be able to change interactively the highlighted products it is needed to
separate it from the creation/recreation of the OpenCSGRenderer.
Setting the highlighted products is not done using the setHighlight method.

highlightsProducts is renamed into compileTimeHightlightedProducts to make it clear
that these highlight come from the source code.