Add namespace lookup to the SourceFile
SourceFile is almost the root of lexical environments (its parent becomes the
builtins). Whenever we encounter a namespace during parsing, we want
them to reuse the same lexical environment. This
allows a namespace's definitions to be placed in multiple files.

There's always a SourceFile, it's the only thing we'd want as the
parent of a namespace, so it's the logical place to hold the
namespaces' scopes.