Switch LocalScope from mix of stuff to std::shared_ptr
The code is grabbing a pointer from a unique_ptr and then fobbing it
off random places; that's definitely not the right intent!

This standardizes on shared_ptr everywhere for scopes.

This commit is missing some changes to parser.y because I
have mixed changes for namespace in there I didn't want to break
out. The changes are straightforward to pull out if this becomes a
separate PR.