Add using and namespace to LocalScope + template-specialized lookup
Rather than repeat the pattern of other classes delving in the
internals of LocalScope to find functions and methods, wrote a
template-specialized lookup function which I'll use in future
commits. Progress on #6134

Add support functions for keeping track of `using` and `namespace`
statements just like assignments, functions, and modules have.

Moved some members which weren't used outside the class to `private`
block. They also should be replaced with the more-generic facility
written in this commit, but trying to minimize changes.