Store, evaluate, and lookup from namespaces
Add templated `lookup_as_namespace` method to Context, which looks in
defined functions/modules for a name, then check assignments for
function (future: module) literals, and does NOT check namespaces
included by `using`.

Namespace contexts belong to the session and all Context have access
to the session, so the context lookup methods are placed in
EvaluationSession.

All the new methods are templated instead of having different names
for each environment and duplicating much code.

The source file's context is available as the namespace "top_level"
and the builtins are, believe it or not, available as the namespace
"builtins".

The templated methods which aren't specialized must be explicitly
instantiated because of separate compilation units.

The namespaces only have access to themself, builtins, and any
namespaces which first were named before themself when evaluating
assignments.

It's possible namespaces declared in files that are `use<>`d instead
of included don't work.