Add namespace::name syntax
Lookup and ModuleInstantiation need to know if they're referencing a
name from a namespace.

Parser and Lexer updated. Unknown question whether to update code to
fail in the parser if attempting to pull a special value from a
namespace. Currently it's just never found, unless maybe it's a
builtin. Whatever way, TBD most desireable way to fail and to make
sure it is done.

Not sure if the "getPrintableName()" should be namespaced? But I
renamed and changed the method so it's only used for the purpose of
printing, so if refactoring needed it's just one place. The other
purpose was checking if the module was "children" -- special case now
moved into a dedicated function.

Pull cloning into ModuleInstantiation because it should own
duplicating itself.