Reduce header pollution by moving includes to cc from headers
A problem all started when I couldn't dynamic_cast a BuiltinContext
and apparently header pollution can be the cause or trigger. These are
the ones that aren't mixed in with other changes I have made. Several
cc files had to be updated because of indirect includes.

Though it turned out it was a typo and I was comparing the
value-to-be-assigned for nullptr and not the
value-from-dynamic-cast. C'est la vie.