Add use<> and allow (but ignore) module instantiations within namespaces
Refactor: BuiltinContext is grabbed magically when first pushed on
the frame stack. This then allows ScopeContext to semantically request
the BuiltinContext when processing a use<> statement instead of just
assuming the parent context is the builtin (because it's not anymore
inside a namespace evaluation).

Refactor: LocalScope's assignments are now private and have an
accessor.

Refactor: parser.y's scope_stack had to be changed to a vector because
you can't iterate through a stack.

use<> within a namespace follows the same semantics as use<> elsewhere
in the program and is the same code, refactored so the existing code
uses the namespace "top_level".