Reduce header pollution (#6164)
* Sort includes into (mostly) alphabetical order

I thought I had to do add something, but it turns out just to be my
mistake. Might as well share this cleanup though.

* Reduce header pollution by moving includes to cc from headers

A problem all started when I couldn't dynamic_cast a BuiltinContext
and apparently header pollution can be the cause or trigger. This is
just a good start toward cleaning up.  Several
cc files had to be updated because of indirect includes.