python: Fully destroy the intepreter on re-evaluation
I have seen a crash in the current code when importing a file
that is not part of a package, like:

```
├── foo
│   ├── bar
│   │   └── baz.py
│   └── __init__.py
└── scad.py
```

scad.py:

```
import os
import sys

if os.getcwd() not in sys.path:
    sys.path.append(os.getcwd())

import foo.bar.baz
```

Would crash `openscad` on re-evaluation of the `scad.py` file.

The crashdump is in `initPython`, somewhere where it tries to empty the `sys.modules` dict:

```
#0  0x00007fa843b7e12e PyUnicode_AsEncodedString (libpython3.13.so.1.0 + 0x17e12e)
#1  0x0000000000d2f093 _Z10initPythonRKNSt7__cxx1112basic_stringIcSt11char_traitsIcESaIcEEEd (/home/damz/Downloads/openscad/build/openscad + 0x92f0>
#2  0x0000000000f0b40b _ZN10MainWindow13parseDocumentEP15EditorInterface (/home/damz/Downloads/openscad/build/openscad + 0xb0b40b)
```

Instead of trying to manually reinit the environment, which is messy and doesn't
reverts all the side-effects that could have happened (like modifying `sys.path`
above), tear down the interpreter with `Py_FinalizeEx` (available since Python 3.6,
released in 2016).