Add delta measurements and fix non-clearing of Measurement
No matter how you close the pop-up menu, it now resets state and
clears the selected lines/points.

You can now click on things in the pop-up menu to copy them to the
clipboard. Previously there was no way besides writing down the
number yourself.

In addition to straight distance, for point-to-point distance
measurement it now shows the delta for each axis.

For point to line segment, it now also shows deltas for each endpoint
plus the deltas for the (imaginary) perpendicular line segment if
the (original) line segment was a line (i.e. infinitely long).

Alphabetize and dedup some headers in MainWindow

Add some comments asking for clarification of confusing things.